ai_api/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
matplotlib==3.10.1
pandas==2.2.3
pillow==11.2.1
schedule==1.2.2
ultralytics==8.3.107
